feat: connect group detail API to backend - Replace localStorage with groupApi.getGroupById() - Simplify group detail page to focus on API integration - Display group info: name, description, invite code, creator - Display members list with avatars - Fix API response type for getGroupById to match backend - Backup old complex implementation as page_old_backup.tsx
